requirements.txt
azure-ai-projects 
azure-identity 
semantic-kernel[azure] 
aiohttp 
python-dotenv
streamlit
matplotlib
plotly
requests
# for streamlit
streamlit>=1.22
fastapi>=0.85
uvicorn>=0.19
python-dotenv>=0.19
azure-identity>=1.12
azure-ai-ml>=1.0
semantic-kernel>=0.3
requests>=2.28
pillow>=9.0
pydantic>=1.10